Fix file directory customer,event,news,securitynotices
