+++ SangmyungBank/src/main/java/com/EmperorPenguin/SangmyungBank/api/customer/domain/repository/CustomerRepository.java
@@ -1,0 +1,6 @@
+package com.EmperorPenguin.SangmyungBank.api.customer.domain.repository;
+
+import com.EmperorPenguin.SangmyungBank.api.customer.domain.customer.Customer;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CustomerRepository extends JpaRepository<Customer, Long> { }
