--- SangmyungBank/src/main/java/com/EmperorPenguin/SangmyungBank/api/event/repository/EventRepository.java
@@ -1,6 +1,0 @@
-package com.EmperorPenguin.SangmyungBank.api.event.repository;
-
-import com.EmperorPenguin.SangmyungBank.api.event.model.Event;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface EventRepository extends JpaRepository<Event, Long> { }
