+++ SangmyungBank/src/main/java/com/EmperorPenguin/SangmyungBank/api/news/domain/repository/NewsRepository.java
@@ -1,0 +1,6 @@
+package com.EmperorPenguin.SangmyungBank.api.news.domain.repository;
+
+import com.EmperorPenguin.SangmyungBank.api.news.domain.news.News;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface NewsRepository extends JpaRepository<News, Long> { }
