+++ SangmyungBank/src/main/java/com/EmperorPenguin/SangmyungBank/api/news/service/NewsService.java
@@ -1,16 +1,13 @@
 package com.EmperorPenguin.SangmyungBank.api.news.service;
 
-import com.EmperorPenguin.SangmyungBank.api.news.model.News;
-import com.EmperorPenguin.SangmyungBank.api.news.repository.NewsRepository;
+import com.EmperorPenguin.SangmyungBank.api.news.domain.news.News;
+import com.EmperorPenguin.SangmyungBank.api.news.domain.repository.NewsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 @Service
 public class NewsService {
