+++ SangmyungBank/src/main/java/com/EmperorPenguin/SangmyungBank/api/securitynotices/domain/repository/SecurityNoticesRepository.java
@@ -1,0 +1,6 @@
+package com.EmperorPenguin.SangmyungBank.api.securitynotices.domain.repository;
+
+import com.EmperorPenguin.SangmyungBank.api.securitynotices.domain.securitynotices.SecurityNotices;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SecurityNoticesRepository extends JpaRepository<SecurityNotices, Long> { }
