+++ SangmyungBank/src/main/java/com/EmperorPenguin/SangmyungBank/api/securitynotices/service/SecurityNoticesService.java
@@ -1,7 +1,7 @@
 package com.EmperorPenguin.SangmyungBank.api.securitynotices.service;
 
-import com.EmperorPenguin.SangmyungBank.api.securitynotices.model.SecurityNotices;
-import com.EmperorPenguin.SangmyungBank.api.securitynotices.repository.SecurityNoticesRepository;
+import com.EmperorPenguin.SangmyungBank.api.securitynotices.domain.securitynotices.SecurityNotices;
+import com.EmperorPenguin.SangmyungBank.api.securitynotices.domain.repository.SecurityNoticesRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
